- Fixing product list and optimize cart api

+++ src/main/kotlin/com/piashcse/controller/CartController.kt
@@ -7,6 +7,8 @@ import com.piashcse.entities.product.ProductTable
 import com.piashcse.entities.user.UserTable
 import com.piashcse.models.PagingData
 import com.piashcse.models.cart.*
+import com.piashcse.utils.extension.alreadyExistException
+import com.piashcse.utils.extension.isNotExistException
 import org.jetbrains.exposed.dao.id.EntityID
 import org.jetbrains.exposed.sql.and
 import org.jetbrains.exposed.sql.transactions.transaction
@@ -16,9 +18,9 @@ class CartController {
         val isProductExist =
             CartItemEntity.find { CartItemTable.userId eq userId and (CartItemTable.productId eq addCart.productId) }
                 .toList().singleOrNull()
-        isProductExist?.apply {
-            this.quantity = this.quantity + addCart.quantity
-        }?.cartResponse() ?: CartItemEntity.new {
+        isProductExist?.let {
+            addCart.productId.alreadyExistException()
+        } ?: CartItemEntity.new {
             this.userId = EntityID(userId, CartItemTable)
             productId = EntityID(addCart.productId, CartItemTable)
             quantity = addCart.quantity
@@ -27,8 +29,8 @@ class CartController {
 
     fun getCartItems(userId: String, pagingData: PagingData) = transaction {
         CartItemEntity.find { CartItemTable.userId eq userId }.limit(pagingData.limit, pagingData.offset).map {
-                it.cartResponse(ProductEntity.find { ProductTable.id eq it.productId }.first().response())
-            }
+            it.cartResponse(ProductEntity.find { ProductTable.id eq it.productId }.first().response())
+        }
     }
 
     fun updateCartQuantity(userId: String, updateCart: UpdateCart) = transaction {
@@ -54,8 +56,14 @@ class CartController {
     }
 
     fun deleteAllFromCart(userId: String) = transaction {
-        CartItemEntity.find { CartItemTable.userId eq userId }.toList().forEach {
-            it.delete()
+        val isEmpty = CartItemEntity.find { CartItemTable.userId eq userId }.toList()
+        if (isEmpty.isEmpty()) {
+            "".isNotExistException()
+        } else {
+            isEmpty.forEach {
+                it.delete()
+            }
+            true
         }
     }
 }
